third_party/blink/web_tests/svg/dynamic-updates/SVGFEConvolveMatrixElement-svgdom-kernelUnitLength-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'kernelUnitLength' property of the SVGFEConvolveMatrixElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS convolveMatrixElement.kernelUnitLengthX.baseVal is 1
PASS convolveMatrixElement.kernelUnitLengthY.baseVal is 1
PASS Math.round(convolveMatrixElement.kernelUnitLengthX.baseVal * 1000) is 50
PASS Math.round(convolveMatrixElement.kernelUnitLengthY.baseVal * 1000) is 50
PASS successfullyParsed is true

TEST COMPLETE
